+++ src/log/xml_utils.cpp
