+++ src/log/xml_utils.h
