+++ src/structures/map.cpp
@@ -1,4 +1,4 @@
-#include "xml_utils.h"
+#include "../log/xml_utils.h"
 #include "map.h"
 
 namespace {
