+++ src/structures/search_result.h
